WFLY-15704 Move ScheduleExpressionTypeUtil#getType method to ScheduleExpressionType class.
